feat: Improve error handling and add new error types (evaluator)
This commit significantly improves error handling in the
lexer and parser.  A `SourceLine` field is added to the
`Error` struct, along with new error types and helper
functions.  The `tokenToSymbol` function and error
messages are enhanced, and type compatibility checks are
improved.  A `WithSourceLine` method simplifies source
line integration.

Additionally, a `WhileStatement` evaluator and
`GetEnvironment` method are added.  Minor code cleanup
and example test files are included.